pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "OctoCarvera"
version = "0.5.13"
description = "OctoPrint integration for the Carvera Air CNC machine over USB serial"
authors = [
    {name = "Jorgen Steen"}
]
readme = {file = "README.md", content-type = "text/markdown"}
dynamic = [
    "license"
]

requires-python = ">=3.7, <4"

# any additional requirements (besides OctoPrint) should be listed here
dependencies = [
    "pyserial",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = [
    "octocarvera",
    "octocarvera.*"
]

[project.entry-points."octoprint.plugin"]
octocarvera = "octocarvera"

[project.urls]
Homepage = "https://github.com/JorgenSteen/OctoCarvera"

[project.optional-dependencies]
develop = [
    "go-task-bin"
]
